Fix a buffer overflow when spawning static creatures with TotemAI.
The core currently does not support spawning creatures that use TotemAI
as AIName via the creature table.

That's because when a grid is loaded, the core spawns the creatures in
that grid via a call to ObjectGridLoader::Visit(CreatureMapType&), which
delegates the actual spawning to a LoadHelper<Creature>() call. Notice
that the template parameter T in this call is always 'Creature', so when
this call creates the C++ object associated with the creature ('new T'),
it will always create a Creature object, never any derived type. When
initializing the Creature object afterwards, via Creature::LoadFromDB(),
this will also set up the creature's AI via AIM_Initialize(), which
could end up selecting TotemAI as the AI based on the creature's AIName.
This is a problem because Unit::Update() also updates the AI via a call
to AI()->UpdateAI(), and in the case of TotemAI that means calling
TotemAI::UpdateAI(), a call that assumes the associated m_creature is
actually a Totem object rather than just a Creature object! If this is
not the case however, then TotemAI trying to access Totem class members
is basically accessing data out of bounds...

Because the core obviously can't control the data that's present in the
database, let's detect this case and avoid the problem by not loading
such creatures.