AntispamMgr: make the shut down member atomic.
This makes sure it's properly synchronized between threads. The member
is continuously read from the AntispamMgr's worker loop, which runs in
its own thread, while it is set in the AntispamMgr's destructor, which
gets called when the AntispamMgr's singleton instance is destroyed,
typically at the end of the main thread.