SqlStmtParameters: cleanup special member functions.
The private-but-not-implemented copy constructor is a relic from the
past to make the class non-copyable, but these days we can just make it
public and deleted. That said, I don't really see a compelling reason to
make this non-copyable.

I also see no reason to explicitly add a default destructor given this
class is not a base class. That way, it nicely follows the so called
"rule of zero" (or the C++ core guideline "if you can avoid defining
default operations, do").

This also avoids a bunch of [-Wdeprecated-copy] GCC/clang compiler
warnings when using recent versions, because an implicitly declared copy
constructor (which is actually used, for example in SqlStatement's copy
constructor) is deprecated when you have the copy assignment declared.