world_map_kalimdor: add missing initializer creating a GhostOPlasmEvent.
Practically speaking this makes no difference at all: for a non-union
aggregate each element that is not an explicitly initialized element,
has no default member initializer, and is not a reference, the element
is copy-initialized from an empty initializer list anyway. But it fixes
a GCC/clang compiler waring: 'warning: missing initializer for member
'GhostOPlasmEvent::summonedMagrami' [-Wmissing-field-initializers]'.